backend/requirements.txt
Flask
Flask_cors
flask_sqlalchemy
Flask-JWT-Extended
google-auth
google-auth-oauthlib
requests